A pseudonym archive

An Ark object can create and remember pseudonyms. Given the same input, it will always return the same pseudonym. No pseudonym will repeat.

Public fields

log

Hashtable for all used pseudonyms. Inputs (keys) are stored as hashes.

parts

Words that will be combined to form pseudonyms.

max_length

Maximum number of possible pseudonyms in the Ark.

alliterate

Logical, generate alliterations by default?

index_allit

indices of alliterations

index_perm

a random permutation of the index

Active bindings

parts

Words that will be combined to form pseudonyms.

max_length

Maximum number of possible pseudonyms in the Ark.

alliterate

Logical, generate alliterations by default?

index_allit

indices of alliterations

index_perm

a random permutation of the index

Methods

Public methods

Ark$new()

Ark$pseudonymize()

Ark$print()

Ark$length()

Ark$clone()

Method new()

Create new ark object.

Usage

Ark$new(alliterate = FALSE)

Arguments

alliterate

Logical. Should the Ark return alliterations by default?

Returns

A new Ark object.

Method pseudonymize()

Create Pseudonyms for input.

Usage

Ark$pseudonymize(..., .alliterate = NULL)

Arguments

...

One or more R objects.

.alliterate

Logical. Return only pseudonyms that are alliterations. Defaults to TRUE if the Ark was created with Ark$new(alliterate = TRUE), FALSE otherwise. If FALSE, pseudonyms may still be alliterations by coincidence.

Returns

Character vector of pseudonyms with same length as input.

Method print()

Pretty-print an Ark object.

Usage

Ark$print(n = NULL)

Arguments

n

A positive integer. The number of example pseudonyms to print.

Method length()

Number of used pseudonyms in an Ark.

Usage

Ark$length()

Method clone()

The objects of this class are cloneable with this method.

Usage

Ark$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.