Add column with pseudonyms to a data frame

Add column with pseudonyms to a data frame

add_pseudonyms(
  .data,
  ...,
  .name = "pseudonym",
  .before = NULL,
  .after = NULL,
  .ark = NULL
)

Arguments

.data
A data frame to add pseudonyms to.

...
Columns to use as keys on which pseudonyms should be based. Supports tidy select. If empty, all columns will be used.

.name
Name of the new column as string.

.before
<tidy-select> Destination of columns selected by .... Supplying neither will move columns to the left-hand side; specifying both is an error.

.after
<tidy-select> Destination of columns selected by .... Supplying neither will move columns to the left-hand side; specifying both is an error.

.ark
An Ark object. If NULL (default) a new Ark is created. Using an existing Ark makes sure that the same input returns the same pseudonym.

Value

A data frame with an additional column containing the pseudonyms.

Examples

add_pseudonyms(mtcars)

#>                                pseudonym  mpg cyl  disp  hp drat    wt  qsec vs
#> Mazda RX4                  Careful Raven 21.0   6 160.0 110 3.90 2.620 16.46  0
#> Mazda RX4 Wag            Elegant Leopard 21.0   6 160.0 110 3.90 2.875 17.02  0
#> Datsun 710              Plastic Pinniped 22.8   4 108.0  93 3.85 2.320 18.61  1
#> Hornet 4 Drive         Annoying Wildfowl 21.4   6 258.0 110 3.08 3.215 19.44  1
#> Hornet Sportabout           Breeze Shrew 18.7   8 360.0 175 3.15 3.440 17.02  0
#> Valiant                    Cheap Peafowl 18.1   6 225.0 105 2.76 3.460 20.22  1
#> Duster 360                Natural Weasel 14.3   8 360.0 245 3.21 3.570 15.84  0
#> Merc 240D            Tender Voalavoanala 24.4   4 146.7  62 3.69 3.190 20.00  1
#> Merc 230              Far-Flung Flamingo 22.8   4 140.8  95 3.92 3.150 22.90  1
#> Merc 280                   Drab Silkworm 19.2   6 167.6 123 3.92 3.440 18.30  1
#> Merc 280C             Reflective Quetzal 17.8   6 167.6 123 3.92 3.440 18.90  1
#> Merc 450SE             Absorbed Marmoset 16.4   8 275.8 180 3.07 4.070 17.40  0
#> Merc 450SL                   Melted Duck 17.3   8 275.8 180 3.07 3.730 17.60  0
#> Merc 450SLC               Flimsy Whippet 15.2   8 275.8 180 3.07 3.780 18.00  0
#> Cadillac Fleetwood     Disgusted Meerkat 10.4   8 472.0 205 2.93 5.250 17.98  0
#> Lincoln Continental            Nice Crow 10.4   8 460.0 215 3.00 5.424 17.82  0
#> Chrysler Imperial        Goofy Centipede 14.7   8 440.0 230 3.23 5.345 17.42  0
#> Fiat 128                    Guarded Goat 32.4   4  78.7  66 4.08 2.200 19.47  1
#> Honda Civic                Frantic Takin 30.4   4  75.7  52 4.93 1.615 18.52  1
#> Toyota Corolla               Kind Thrush 33.9   4  71.1  65 4.22 1.835 19.90  1
#> Toyota Corona                 Curvy Flea 21.5   4 120.1  97 3.70 2.465 20.01  1
#> Dodge Challenger         Clumsy Pheasant 15.5   8 318.0 150 2.76 3.520 16.87  0
#> AMC Javelin               Helpful Badger 15.2   8 304.0 150 3.15 3.435 17.30  0
#> Camaro Z28                  Icky Tarsier 13.3   8 350.0 245 3.73 3.840 15.41  0
#> Pontiac Firebird               Hard Guan 19.2   8 400.0 175 3.08 3.845 17.05  0
#> Fiat X1-9                       Wiry Asp 27.3   4  79.0  66 4.08 1.935 18.90  1
#> Porsche 914-2       Scandalous Wolverine 26.0   4 120.3  91 4.43 2.140 16.70  0
#> Lotus Europa               Omniscient Ox 30.4   4  95.1 113 3.77 1.513 16.90  1
#> Ford Pantera L          Possessive Bison 15.8   8 351.0 264 4.22 3.170 14.50  0
#> Ferrari Dino          Functional Lobster 19.7   6 145.0 175 3.62 2.770 15.50  0
#> Maserati Bora          Separate Lungfish 15.0   8 301.0 335 3.54 3.570 14.60  0
#> Volvo 142E              Marked Chickadee 21.4   4 121.0 109 4.11 2.780 18.60  1
#>                     am gear carb
#> Mazda RX4            1    4    4
#> Mazda RX4 Wag        1    4    4
#> Datsun 710           1    4    1
#> Hornet 4 Drive       0    3    1
#> Hornet Sportabout    0    3    2
#> Valiant              0    3    1
#> Duster 360           0    3    4
#> Merc 240D            0    4    2
#> Merc 230             0    4    2
#> Merc 280             0    4    4
#> Merc 280C            0    4    4
#> Merc 450SE           0    3    3
#> Merc 450SL           0    3    3
#> Merc 450SLC          0    3    3
#> Cadillac Fleetwood   0    3    4
#> Lincoln Continental  0    3    4
#> Chrysler Imperial    0    3    4
#> Fiat 128             1    4    1
#> Honda Civic          1    4    2
#> Toyota Corolla       1    4    1
#> Toyota Corona        0    3    1
#> Dodge Challenger     0    3    2
#> AMC Javelin          0    3    2
#> Camaro Z28           0    3    4
#> Pontiac Firebird     0    3    2
#> Fiat X1-9            1    4    1
#> Porsche 914-2        1    5    2
#> Lotus Europa         1    5    2
#> Ford Pantera L       1    5    4
#> Ferrari Dino         1    5    6
#> Maserati Bora        1    5    8
#> Volvo 142E           1    4    2